TextInput case rename pt 2
